refuge_requirements.txt
streamlit
opencv-python
pillow
requests
numpy
